backend/ipment/apps/Configuration/serializers.py
from rest_framework import serializers
from .models import MeasurementUnit, Label

class MeasurementUnitSerializer(serializers.ModelSerializer):
    class Meta:
        model = MeasurementUnit
        fields = '__all__'
        
class LabelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Label
        fields = '__all__'
