backend/ipment/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
python-dotenv
python-decouple
requests